Authization usig Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Question;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\AskQuestionRequest;
 
 
@@ -67,6 +68,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(Gate::denies('update-question',$question)){
+
+            abort(403,'Access denied!');
+        }
         return view('questions.edit',compact('question'));
     }
 
@@ -93,6 +98,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(Gate::denies('delete-question',$question)){
+
+            abort(403,'Access denied!');
+        }
         $question->delete();
 
         return redirect()->route('questions.index')->with('success','you question has been deleted');
